requirements.txt
Flask
pymongo
Flask-Bcrypt
cohere
python-dotenv
fpdf
gunicorn